Add files via upload
Number Guessing game.

How it works:
the Random module is imported. It is used to create the random number, that you'll be attempting to guess.
A for loop is used and the range is 1,6, so you get 5 guesses.
If your guess is correct, it will print a small message, and exit() the program. It will either give you a couple more guesses (if you haven't used all your 5 guesses, or stop. the program.
If your guess is larger than the random number, it will print a small message, and either give you a couple more guesses (if you haven't used all your 5 guesses, or stop. the program
If your guess is smaller than the random number, it will print a small message, and either give you a couple more guesses (if you haven't used all your 5 guesses, or stop. the program.