Add files via upload
Calculator.

How it works:
The user is asked to input two integers (x, y), then is asked to input the operator (what do they want to do, subtract, add, and etc).
The Addition, Subtraction, Multiplication, and Division functions are defined.
If statements are used to determine what the user wanted, so if the earlier operator input was *, it will know what to do. For this case, it'll use the multiplication function.